etl/requirements.txt
requests==2.31.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pandas==2.1.4
tenacity==8.2.3
pydantic==2.5.3
pydantic-settings==2.1.0
